% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.r
\name{fit_model}
\alias{fit_model}
\title{Fit stan model}
\usage{
fit_model(timeseries, shapes, timestep = 1, period_and_lag = c(5, 7),
  identifier = "ADM2_NAME", popid = "totpop2019", interaction = c(1),
  distrib = 0, model_path = system.file("extdata/template.stan",
  package = "EpiCastR"), fit_meth = "vb")
}
\description{
Fit stan model
}
