% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forecast.r
\name{make_forecast}
\alias{make_forecast}
\title{make_forecast}
\usage{
make_forecast(
  timeseries_mat,
  shapes,
  identifier = "ADM2_NAME",
  day_of_forecast = NULL,
  do_score_forecast = TRUE,
  fit_over = 1000,
  timestep = 1,
  period_and_lag = c(5, 7),
  interaction = c(1),
  distrib = 1,
  fit_meth = "vb",
  chains = 1,
  iter = 100,
  warmup = 50,
  cores = 1,
  timehorizons = c(7, 14, 28),
  thresholds = c(1, 2, 5, 10, 20),
  close_down = FALSE,
  con_mat = NULL,
  buffer = 0
)
}
\arguments{
\item{timeseries_mat}{a matrix of timeseries of cases for each geographical region [region x timestep]}

\item{shapes}{sf or sp object with polygon geometry of regions and population size}

\item{day_of_forecast}{day of outbreak to make forecast - default last day of data}

\item{fit_over}{number of timesteps to fit over.}

\item{timestep}{timestep to use in fit (relative to the "timeseries"" timestep)}

\item{period_and_lag}{vector of 2 values in timesteps: first the period of influencial cases, second the lag between influential cases and fitted data}

\item{interaction}{vector with numbers to indicate the types of interaction to include in the model: 1. Gravity model. 2. Gravity model with population density. 3. Power law (no population info). 4. Adjacency model.}

\item{distrib}{poisson = 0, negative binomial = 1}

\item{fit_meth}{variational bayes 'vb', Hamiltonian MC 'nuts'}

\item{chains}{number of MC chains}

\item{iter}{number of iterations per chain}

\item{warmup}{number of iterations in warmup phase}

\item{cores}{number of cores to use}
}
\description{
make_forecast
}
