% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_historical_ha.r
\name{run_historical_ha}
\alias{run_historical_ha}
\title{Run historical ha}
\usage{
run_historical_ha(days, fit_over, iters = 10, warmup = 5, cores = 1,
  chains = 1)
}
\description{
Run historical ha
}
