% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_forecasts.r
\name{score_forecasts}
\alias{score_forecasts}
\title{Score forecasts}
\usage{
score_forecasts(risk_values, diff_cases, timehorizon, day_of_forecast)
}
\arguments{
\item{risk_values}{tbc}

\item{diff_cases}{tbc}

\item{timehorizon}{tbc}

\item{day_of_forecast}{tbc}
}
\description{
Score forecasts
}
