% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_parameters.r
\name{prepare_stan_inputs}
\alias{prepare_stan_inputs}
\title{Prepare stan inputs}
\usage{
prepare_stan_inputs(timeseries, shapes, timestep = 1,
  period_and_lag = c(5, 7), identifier = "ADM2_NAME",
  popid = "totpop2019", interaction = c(1), distrib = 0)
}
\description{
Prepare stan inputs
}
